Make inventory loading on join

+++ pkg/game/items/slot.go
@@ -2,7 +2,7 @@ package items
 
 type Slot struct {
 	IsPresent bool
-	ItemID    int32
-	ItemCount int8
+	ItemID    int
+	ItemCount int
 	NBT       map[string]string // DEBT not clear atm how to handle this NBT compound
 }
